Fix two metacluster issues:
1. When retrying the transaction to register a restoring cluster, don't choose a new ID if the current ID matches the one recorded for the restoring cluster
2. A metacluster test was incorrectly handling the case where a transaction was retried with unknown result and had committed successfully

+++ fdbserver/workloads/MetaclusterManagementWorkload.actor.cpp
@@ -358,8 +358,9 @@ struct MetaclusterManagementWorkload : TestWorkload {
 
 				state std::vector<Future<Void>> deleteFutures;
 				for (auto const& f : getFutures) {
-					ASSERT(f.get().present());
-					deleteFutures.push_back(TenantAPI::deleteTenantTransaction(tr, f.get().get()));
+					if (f.get().present()) {
+						deleteFutures.push_back(TenantAPI::deleteTenantTransaction(tr, f.get().get()));
+					}
 				}
 
 				wait(waitForAll(deleteFutures));
